Dedup SupervisorApi
