+++ crates/node/rpc/src/jsonrpsee.rs
@@ -2,11 +2,9 @@
 
 use crate::{OutputResponse, ProtocolVersion, SafeHeadResponse, SuperchainSignal};
 use alloy_eips::BlockNumberOrTag;
-use alloy_primitives::B256;
 use core::net::IpAddr;
 use jsonrpsee::{core::RpcResult, proc_macros::rpc};
 use kona_genesis::RollupConfig;
-use kona_interop::{ExecutingDescriptor, SafetyLevel};
 use kona_p2p::{PeerCount, PeerDump, PeerInfo, PeerStats};
 use kona_protocol::SyncStatus;
 
@@ -143,34 +141,4 @@ pub trait EngineApiExt {
     /// See: <https://specs.optimism.io/protocol/exec-engine.html#engine_signalsuperchainv1>
     #[method(name = "signalSuperchainV1")]
     async fn signal_superchain_v1(&self, signal: SuperchainSignal) -> RpcResult<ProtocolVersion>;
-}
-
-/// Supervisor API for interop.
-#[cfg_attr(not(feature = "client"), rpc(server, namespace = "supervisor"))]
-#[cfg_attr(feature = "client", rpc(server, client, namespace = "supervisor"))]
-pub trait SupervisorApi {
-    /// Checks if the given inbox entries meet the given minimum safety level.
-    #[method(name = "checkAccessList")]
-    async fn check_access_list(
-        &self,
-        inbox_entries: Vec<B256>,
-        min_safety: SafetyLevel,
-        executing_descriptor: ExecutingDescriptor,
-    ) -> RpcResult<()>;
-}
-
-#[cfg(feature = "client")]
-impl<T> crate::CheckAccessList for T
-where
-    T: SupervisorApiClient + Send + Sync,
-{
-    async fn check_access_list(
-        &self,
-        inbox_entries: &[B256],
-        min_safety: SafetyLevel,
-        executing_descriptor: ExecutingDescriptor,
-    ) -> Result<(), crate::InteropTxValidatorError> {
-        Ok(T::check_access_list(self, inbox_entries.to_vec(), min_safety, executing_descriptor)
-            .await?)
-    }
 }
