+++ crates/supervisor/rpc/Cargo.toml
