+++ crates/supervisor/rpc/src/lib.rs
@@ -1,7 +1,7 @@
 #![doc = include_str!("../README.md")]
 
 #[cfg(feature = "jsonrpsee")]
-mod jsonrpsee;
+pub mod jsonrpsee;
 #[cfg(all(feature = "jsonrpsee", feature = "client"))]
 pub use jsonrpsee::SupervisorApiClient;
 #[cfg(feature = "jsonrpsee")]
